add afdb-dimer
