+++ configs/dataset/distill-heterodimer.yaml
@@ -1,0 +1,23 @@
+_yaml_: "default-train.yaml"
+
+name: AFDB-heterodimer
+type: "heterodimer-distillation"
+prob_drop_apo: 0.0
+prob_drop_struct_token: 0.0
+
+sampler:
+  _registry_: "data_sampler"
+  _class_: UniformSampler
+  level: interface
+
+cropper:
+  _registry_: "data_cropper"
+  _class_: AlphaFold3Cropper
+  w_contiguous: 0.2
+  w_spatial: 0.4
+  w_spatial_interface: 0.4
+
+prob_perturbation: 1.0
+apo_perturb:
+  bioprior:
+    max_steps: 10
