+++ src/kfold/training/dataset/datasets/__init__.py
@@ -2,8 +2,8 @@ from .train_dataset import TrainingDataset, TrainingDatasetConfig  # noqa
 from .val_dataset import ValidationDataset, ValidationDatasetConfig
 from .rcsb import DisorderedPDBTrainingDataset, RCSBTrainingDataset
 from .distillation import DistillationDataset
-from .monomer import MonomerDistillationDataset, RNAMonomerDistillationDataset
-from .homodimer import HomodimerDistillationDataset
+from .monomer import ProteinMonomerDistillationDataset, RNAMonomerDistillationDataset
+from .dimer import HeterodimerDistillationDataset, HomodimerDistillationDataset
 
 
 def get_training_dataset_cls(
@@ -16,11 +16,13 @@ def get_training_dataset_cls(
             return DisorderedPDBTrainingDataset
         case "distillation":
             return DistillationDataset
-        case "monomer-distillation":
-            return MonomerDistillationDataset
+        case "protein-monomer-distillation":
+            return ProteinMonomerDistillationDataset
         case "rna-monomer-distillation":
             return RNAMonomerDistillationDataset
         case "homodimer-distillation":
             return HomodimerDistillationDataset
+        case "heterodimer-distillation":
+            return HeterodimerDistillationDataset
         case _:
             raise ValueError(f"Unsupported training dataset type: {train_config.type}")
